Refactor ModelList in new model_list folder
