+++ web/src/components/hugging_face/model_list/ModelDisplay.tsx
@@ -1,0 +1,107 @@
+import React from "react";
+import { Grid, List, Typography } from "@mui/material";
+import ModelCard from "../ModelCard";
+import ModelListItem from "./ModelListItem";
+import { UnifiedModel } from "../../../stores/ApiTypes";
+import { ModelSource } from "./useModels";
+import { useModelDownloadStore } from "../../../stores/ModelDownloadStore";
+
+type ViewMode = "grid" | "list";
+
+interface ModelDisplayProps {
+  models: UnifiedModel[];
+  viewMode: ViewMode;
+  modelSource: ModelSource;
+  modelSearchTerm: string;
+  handleDeleteClick: (modelId: string) => void;
+  handleShowInExplorer: (modelId: string) => void;
+  ollamaBasePath: string | null | undefined;
+}
+
+const ModelDisplay: React.FC<ModelDisplayProps> = ({
+  models,
+  viewMode,
+  modelSource,
+  modelSearchTerm,
+  handleDeleteClick,
+  handleShowInExplorer,
+  ollamaBasePath
+}) => {
+  const downloadStore = useModelDownloadStore();
+
+  const startDownload = React.useCallback(
+    (model: UnifiedModel) => {
+      const repoId = model.repo_id || model.id;
+      downloadStore.startDownload(
+        repoId,
+        model.type ?? "",
+        model.path ?? undefined
+      );
+      downloadStore.openDialog();
+    },
+    [downloadStore]
+  );
+
+  if (models.length === 0) {
+    return (
+      <Typography variant="body1" sx={{ mt: 2 }}>
+        {modelSearchTerm ? `No models found for "${modelSearchTerm}"` : ""}
+      </Typography>
+    );
+  }
+
+  const allowDelete = modelSource === "downloaded";
+  const allowDownload = modelSource === "recommended";
+  const allowShowInExplorer = modelSource === "downloaded";
+
+  if (viewMode === "grid") {
+    return (
+      <Grid className="model-grid-container" container spacing={3}>
+        {models.map((model: UnifiedModel, idx: number) => (
+          <Grid
+            className="model-grid-item"
+            item
+            xs={12}
+            sm={12}
+            md={6}
+            lg={4}
+            xl={3}
+            key={`${model.id}-${idx}`}
+          >
+            <ModelCard
+              model={model}
+              handleModelDelete={allowDelete ? handleDeleteClick : undefined}
+              onDownload={
+                allowDownload ? () => startDownload(model) : undefined
+              }
+              handleShowInExplorer={
+                allowShowInExplorer ? handleShowInExplorer : undefined
+              }
+              ollamaBasePath={ollamaBasePath}
+            />
+          </Grid>
+        ))}
+      </Grid>
+    );
+  } else {
+    return (
+      <List>
+        {models.map((model: UnifiedModel, idx: number) => (
+          <ModelListItem
+            key={`${model.id}-${idx}`}
+            model={model}
+            handleModelDelete={allowDelete ? handleDeleteClick : undefined}
+            onDownload={allowDownload ? () => startDownload(model) : undefined}
+            handleShowInExplorer={
+              allowShowInExplorer ? handleShowInExplorer : undefined
+            }
+            hideMissingInfo={modelSource === "recommended"}
+            ollamaBasePath={ollamaBasePath}
+          />
+        ))}
+      </List>
+    );
+  }
+};
+
+export default React.memo(ModelDisplay);
