+++ web/src/components/hugging_face/model_list/ModelListItem.styles.ts
